Doc: Adding docker doc, with a new writing method
I use a new method who are most optimized for emacs and especially for
org mode.

In fact, I use more header "*", less words and more concise
explainations.
This result of a bunch of new shorts lines and can be weird to read at
firstly or a bit mess. But I applied same pattern as mindmaps "syntax".
It can maybe be more productive to learn that and retrieve infos.